Response Table Teal Module

This module produces a response summary table that matches the STREAM template rspt01

tm_t_rsp(label, dataname, arm_var = "ARM", arm_var_choices = arm_var,
  arm_ref_comp = NULL, paramcd = "OVRSPI", paramcd_choices = paramcd,
  strata_var = NULL, strata_var_choices = strata_var, pre_output = NULL,
  post_output = NULL, code_data_processing = NULL)

Arguments

label
menue item label of the module in the teal app

dataname
analysis data used in teal module, needs to be available in the list passed to the data argument of init. Note that the data is expected to be in vertical form with the PARAMCD variable filtering to one observation per patient.

arm_var
single name of variable in analysis data that is used as col_by argument for the respective tern function.

arm_var_choices
vector with variable names that can be used as arm_var

arm_ref_comp
optional, if specified it must be a named list with each element corresponding to an arm variable in ASL and the element must be another list with the elements named ref and comp that the defined the default reference and comparison arms when the arm variable is changed.

paramcd
single selected endpoint filtered with PARAMCD variable

paramcd_choices
vector with paramcd choices

strata_var
vector with variable names that should be used for stratification

strata_var_choices
vector with possible choices for strata_var

pre_output
optional, shiny.tag with text placed before the output to put the output into context. For example the helpText elements are useful.

post_output
optional, shiny.tag with text placed before the output to put the output into context. For example the helpText elements are useful.

code_data_processing
string with data preprocessing before the teal app is initialized

Value

an module object

Details

Additional standard UI inputs include responders, ref_arm, comp_arm and combine_arm (default FALSE)

Default values of the inputs var_arm, ref_arm and comp_arm are set to NULL, and updated accordingly based on seletion of paramcd and var_arm

This display order of response categories in partitioned statistics section inherits the factor level order of the source data. Use factor and its levels argument to manipulate the source data in order to include/exclude or re-categorize response categories and arrange the display order. If response categories are "Missing" or "Not Evaluable (NE)" or "Missing or uneavluable", 95% confidence interval will not be calculated.

Reference arms automatically combined if multiple arms selected as reference group.

Examples

 

# NOT RUN {
library(random.cdisc.data)

ASL <- radam('ASL', start_with = list(
  ITTFL = 'Y',
  SEX = c("M", "F"),
  MLIVER = paste("mliver", 1:3),
  ARM = paste("ARM", LETTERS[1:3])
))

ARS <- radam('ARS', ADSL = ASL)

x <- teal::init(
  data = list(ASL = ASL, ARS = ARS),
  modules = root_modules(
    tm_t_rsp(
       label = "Response Table",
       dataname = 'ARS',
       arm_var = "ARM",
       arm_var_choices = c("ARM", "ARMCD"),
       paramcd = "BESRSPI",
       paramcd_choices = unique(ARS$PARAMCD),
       strata_var = "SEX",
       strata_var_choices = c("SEX", "MLIVER")
    )
  )
)

shinyApp(x$ui, x$server)

# }