Summarize Variable Teal Module

Summarize Variable Teal Module

tm_t_summarize_variables(label, dataname, arm_var, arm_var_choices = arm_var,
  summarize_vars, summarize_vars_choices = summarize_vars,
  pre_output = NULL, post_output = NULL, code_data_processing = NULL)

Arguments

label
menue item label of the module in the teal app

dataname
analysis data used in teal module, needs to be available in the list passed to the data argument of init. Note that the data is expected to be in vertical form with the PARAMCD variable filtering to one observation per patient.

arm_var
single name of variable in analysis data that is used as col_by argument for the respective tern function.

arm_var_choices
vector with variable names that can be used as arm_var

summarize_vars
character vector with variable names that are selected by default

summarize_vars_choices
character vector with variable names that can be selected (for summary)

pre_output
optional, shiny.tag with text placed before the output to put the output into context. For example the helpText elements are useful.

post_output
optional, shiny.tag with text placed before the output to put the output into context. For example the helpText elements are useful.

code_data_processing
string with data preprocessing before the teal app is initialized

Examples

# NOT RUN {
library(random.cdisc.data)

ASL <- radam('ASL', N = 1000)

x <- teal::init(
  data = list(ASL = ASL),
  modules = root_modules(
    tm_t_summarize_variables(
       label = "Demographic Table",
       dataname = "ASL",
       arm_var = "ARM",
       arm_var_choices = c("ARM", "ARMCD"),
       summarize_vars =  c("SEX"),
       summarize_vars_choices = c("SEX", "RACE", "BAGE")
   )
  )
)

shinyApp(x$ui, x$server)

# }